Moved send_payment_mail into mail.py. Alos, template used wrong value before.
