+++ stregsystem/mail.py
@@ -4,10 +4,10 @@ import logging
 
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
-from .utils import money
 from django.conf import settings
 from django.template.loader import render_to_string
-
+from django.utils.html import escape
+from stregsystem.templatetags.stregsystem_extras import money
 
 logger = logging.getLogger(__name__)
 
@@ -33,3 +33,29 @@ def send_welcome_mail(member):
 
     msg.attach(MIMEText(html, 'html'))
     send_email(member.email, msg.as_string())
+
+
+def send_payment_mail(member, amount, mobilepay_comment):
+    if hasattr(settings, 'TEST_MODE'):
+        return
+    msg = MIMEMultipart()
+    msg['From'] = 'treo@fklub.dk'
+    msg['To'] = member.email
+    msg['Subject'] = 'Stregsystem payment'
+
+    context = dict()
+    context.update(vars(member))
+    context.update({'formatted_amount': money(amount)})
+    # TODO: not sure if escape serves any purpose here, since it's already being passed through a template.
+    context.update({'mobilepay_comment': escape(mobilepay_comment)})
+
+    target_template = "deposit_manual.html" if mobilepay_comment else "deposit_automatic.html"
+    html = render_to_string(f"mail/{target_template}", context)
+
+    msg.attach(MIMEText(html, 'html'))
+
+    try:
+        smtpObj = smtplib.SMTP('localhost', 25)
+        smtpObj.sendmail('treo@fklub.dk', member.email, msg.as_string())
+    except Exception as e:
+        logger.error(str(e))
