+++ stregsystem/templates/mail/deposit_automatic.html
@@ -3,13 +3,13 @@
     <body>
         <p>
             Hej {{ firstname }}!<br><br>
-            Vi har indsat {{ formatted_balance }} stregdollars på din konto: "{{ username }}". <br><br>
+            Vi har indsat {{ formatted_amount }} stregdollars på din konto: "{{ username }}". <br><br>
             Hvis du ikke ønsker at modtage flere mails som denne kan du skrive en mail til: <a href="mailto:treo@fklub.dk?Subject=Klage" target="_top">treo@fklub.dk</a><br><br>
             Mvh,<br>
             TREOen<br>
             ====================================================================<br>
             Hello {{ firstname }}!<br><br>
-            We've inserted {{ formatted_balance }} stregdollars on your account: "{{ username }}". <br><br>
+            We've inserted {{ formatted_amount }} stregdollars on your account: "{{ username }}". <br><br>
             If you do not desire to receive any more mails of this sort, please file a complaint to: <a href="mailto:treo@fklub.dk?Subject=Klage" target="_top">treo@fklub.dk</a><br><br>
             Kind regards,<br>
             TREOen
