+++ stregsystem/utils.py
@@ -1,10 +1,7 @@
 import logging
 import re
-import smtplib
 
 from django.utils.dateparse import parse_datetime
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.http import HttpResponse
@@ -12,7 +9,6 @@ from django.test.runner import DiscoverRunner
 
 from django.db.models import Count, F, Q, QuerySet
 from django.utils import timezone
-from stregsystem.templatetags.stregsystem_extras import money
 
 import qrcode
 import qrcode.image.svg
@@ -91,37 +87,6 @@ def date_to_midnight(date):
     :return: the date as a timezone aware datetime at midnight
     """
     return timezone.make_aware(timezone.datetime(date.year, date.month, date.day, 0, 0))
-
-
-def send_payment_mail(member, amount, mobilepay_comment):
-    if hasattr(settings, 'TEST_MODE'):
-        return
-    msg = MIMEMultipart()
-    msg['From'] = 'treo@fklub.dk'
-    msg['To'] = member.email
-    msg['Subject'] = 'Stregsystem payment'
-
-    formatted_amount = money(amount)
-
-    from django.utils.html import escape
-    from django.template.loader import render_to_string
-
-    context = dict()
-    context.update(vars(member))
-    context.update({'formatted_balance': money(member.balance)})
-    # TODO: not sure if escape serves any purpose here, since it's already being passed through a template.
-    context.update({'mobilepay_comment': escape(mobilepay_comment)})
-
-    target_template = "deposit_manual.html" if mobilepay_comment else "deposit_automatic.html"
-    html = render_to_string(f"mail/{target_template}", context)
-
-    msg.attach(MIMEText(html, 'html'))
-
-    try:
-        smtpObj = smtplib.SMTP('localhost', 25)
-        smtpObj.sendmail('treo@fklub.dk', member.email, msg.as_string())
-    except Exception as e:
-        logger.error(str(e))
 
 
 def parse_csv_and_create_mobile_payments(csv_file):
